refactor: Create BaseGitHubService and extract shared GitHub API request logic
